% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pglmm.R
\name{probability_estimation}
\alias{probability_estimation}
\title{Probability estimation}
\usage{
probability_estimation(mat, vcv, perspective = "rows", maxit = 40,
  ss = 0.01, tolpql = 10^-6, maxitpql = 200)
}
\value{
The probability matrix
}
\description{
Probability estimation
}

