% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomize}
\alias{randomize}
\title{Randomize unipartite networks}
\usage{
randomize(x, pmat, unilinks_null, bilinks_null, N, type, fprob, degree)
}
\arguments{
\item{x}{A matrix}
}
\description{
Fill the matrix with ones
}
