% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomize_bipartite}
\alias{randomize_bipartite}
\title{Randomize unipartite networks}
\usage{
randomize_bipartite(x, pmat, unilinks_null, N, fprob)
}
\arguments{
\item{x}{A matrix}
}
\description{
Fill the matrix with ones
}

