% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{resoldre}
\alias{resoldre}
\title{Main program}
\usage{
resoldre(mat = NULL, pmat = NULL, cormed = NULL, randomizations = 1,
  verbose = FALSE)
}
\arguments{
\item{cormed}{A correlation-informed matrix used to bias the randomization.}

\item{m}{A matrix to be randomized.}
}
\value{
The result of the randomization in some way...
}
\description{
Informed randomizations main script
}

