% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{resoldre}
\alias{resoldre}
\title{Informed Randomizations}
\usage{
resoldre(mat, pmat = NULL, cormed , randomizations = 1, bipartite = FALSE, seed=NULL, nperm=10^4, perspective="rows", degree=NULL, fprob=0, maxit=40, ss=0.1, tolpql=10^-6, maxitpql=200, f=NULL, ...)
}
\arguments{
\item{mat}{A matrix to be randomized.}

\item{cormed}{A correlation matrix used to bias the randomization.}
}
\value{
The result of the randomization in some way...
}
\description{
This functions performs different type of randomization of an adjacency matrix given either a probability of correlation matrix.
}

